+++ app/src/main/java/woowacourse/shopping/data/payment/dto/AvailableTime.kt
@@ -1,5 +1,4 @@
-package woowacourse.shopping.data.coupon.dto
-
+package woowacourse.shopping.data.payment.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -9,5 +8,5 @@ data class AvailableTime(
     @SerialName("end")
     val end: String,
     @SerialName("start")
-    val start: String
+    val start: String,
 )
