+++ app/src/main/java/woowacourse/shopping/data/payment/dto/CouponResponse.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.data.coupon.dto
+package woowacourse.shopping.data.payment.dto
 
 typealias CouponResponse = List<CouponResponseItem>
