+++ app/src/main/java/woowacourse/shopping/data/payment/dto/CouponResponseItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.coupon.dto
+package woowacourse.shopping.data.payment.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
