+++ app/src/main/java/woowacourse/shopping/data/payment/dto/OrderRequestBody.kt
@@ -1,5 +1,4 @@
-package woowacourse.shopping.data.order.dto
-
+package woowacourse.shopping.data.payment.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -7,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class OrderRequestBody(
     @SerialName("cartItemIds")
-    val cartItemIds: List<Int>
+    val cartItemIds: List<Int>,
 )
