+++ app/src/main/java/woowacourse/shopping/data/payment/repository/CouponRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.payment.repository
 
-import woowacourse.shopping.data.coupon.dto.CouponResponse
 import woowacourse.shopping.data.payment.CouponFetchResult
+import woowacourse.shopping.data.payment.dto.CouponResponse
 
 interface CouponRemoteDataSource {
     suspend fun fetchCoupons(): CouponFetchResult<CouponResponse>
