+++ app/src/main/java/woowacourse/shopping/data/payment/repository/PaymentRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.payment.repository
 
-import woowacourse.shopping.data.coupon.dto.CouponResponse
 import woowacourse.shopping.data.payment.CouponFetchResult
 import woowacourse.shopping.data.payment.OrderRequestResult
+import woowacourse.shopping.data.payment.dto.CouponResponse
 
 interface PaymentRepository {
     suspend fun fetchAllCoupons(): CouponFetchResult<CouponResponse>
