+++ app/src/main/java/woowacourse/shopping/data/util/RetrofitService.kt
@@ -11,10 +11,10 @@ import retrofit2.http.Query
 import woowacourse.shopping.data.carts.dto.CartItemRequest
 import woowacourse.shopping.data.carts.dto.CartQuantity
 import woowacourse.shopping.data.carts.dto.CartResponse
-import woowacourse.shopping.data.coupon.dto.CouponResponse
 import woowacourse.shopping.data.goods.dto.Content
 import woowacourse.shopping.data.goods.dto.GoodsResponse
-import woowacourse.shopping.data.order.dto.OrderRequestBody
+import woowacourse.shopping.data.payment.dto.CouponResponse
+import woowacourse.shopping.data.payment.dto.OrderRequestBody
 
 interface RetrofitService {
     @GET("/products")
