+++ app/src/main/java/woowacourse/shopping/feature/cart/cartdetail/adapter/CartItemSkeletonViewHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.feature.cart.adapter
+package woowacourse.shopping.feature.cart.cartdetail.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartSkeletonBinding
 
-class SkeletonViewHolder(
+class CartItemSkeletonViewHolder(
     val binding: ItemCartSkeletonBinding,
 ) : RecyclerView.ViewHolder(binding.root)
