+++ app/src/main/res/layout/fragment_cart.xml
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
-        tools:context=".feature.cart.CartFragment">
+        tools:context=".feature.cart.cartdetail.CartFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_cart_items"
