+++ app/src/main/res/layout/fragment_order.xml
@@ -13,7 +13,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".feature.cart.OrderFragment">
+        tools:context=".feature.cart.order.OrderFragment">
 
         <androidx.core.widget.NestedScrollView
             android:layout_width="match_parent"
